backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
Pillow==10.1.0
faiss-cpu==1.7.4
numpy==1.24.3
gunicorn==21.2.0
safetensors==0.4.1
git+https://github.com/openai/CLIP.git
scikit-learn==1.3.2
requests==2.31.0
tqdm==4.66.1
